--- Debugger/ILogger.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Debugger
-{
-    public interface ILogger
-    {
-        void Log(string message, Level level = Level.Info, Exception exception = null);
-    }
-}
