--- Models/Models/Account.cs
@@ -1,24 +1,0 @@
-﻿using Newtonsoft.Json;
-using System;
-
-namespace Models.Models
-{
-    [JsonObject]
-    public class Account
-    {
-        [JsonProperty("Name")]
-        public string Name { get; set; }
-
-        [JsonProperty("Password")]
-        public string Password { get; set; }
-
-        [JsonProperty("Email")]
-        public string Email { get; set; }
-
-        [JsonProperty("Id")]
-        public string Id { get; set; }
-
-        [JsonProperty("Guid")]
-        public string Guid { get; set; }
-    }
-}
