+++ Models/Review.cs
@@ -1,0 +1,21 @@
+﻿namespace Models
+{
+    public class Review
+    {
+        public string UniversityId { get; set; }
+
+        public string FacultyId { get; set; }
+
+        public string LecturerId { get; set; }
+
+        public string ProgrammeId { get; set; }
+
+        public string UserId { get; set; }
+
+        public string Text { get; set; }
+
+        public string Value { get; set; }
+
+        public string Id { get; set; }
+    }
+}
