+++ Models/University.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Models
+{
+    public class University
+    {
+        public string Name { get; set; }
+
+        public string Description { get; set; }
+        
+        public string Location { get; set; }
+
+        public DateTime FoundingDate { get; set; }
+
+        public string Id { get; set; }
+    }
+}
