+++ Webserver/Controllers/AccountController.cs
@@ -1,12 +1,12 @@
-﻿using System.Linq;
+﻿using Microsoft.AspNet.Identity;
+using Microsoft.AspNet.Identity.Owin;
+using Microsoft.Owin.Security;
+using System.Linq;
 using System.Threading.Tasks;
 using System.Web;
 using System.Web.Mvc;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin.Security;
 using Webserver.Models;
-using Webserver.ViewModels;
+using Webserver.Models.ViewModels.Account;
 
 namespace Webserver.Controllers
 {
