+++ Webserver/Controllers/ReviewsController.cs
@@ -1,14 +1,14 @@
-﻿using System.Collections.Generic;
+﻿using Models;
+using System.Collections.Generic;
 using System.Net;
 using System.Web.Mvc;
 using Webserver.Data.Repositories;
-using Webserver.Models;
 
 namespace Webserver.Controllers
 {
     public class ReviewsController : Controller
     {
-        private IReviewRepository _repository;
+        private readonly IReviewRepository _repository;
 
         public ReviewsController(IReviewRepository repository)
         {
@@ -24,7 +24,7 @@ namespace Webserver.Controllers
             }
 
             ViewBag.UniversityId = id;
-            IEnumerable<Review> reviews = _repository.GetMany(review => review.UniGuid == id);
+            IEnumerable<Review> reviews = _repository.GetMany(review => review.UniversityId == id);
             return View(reviews);
         }
 
@@ -37,7 +37,7 @@ namespace Webserver.Controllers
             }
 
             ViewBag.FacultyId = id;
-            IEnumerable<Review> reviews = _repository.GetMany(review => review.FacultyGuid == id);
+            IEnumerable<Review> reviews = _repository.GetMany(review => review.FacultyId == id);
             return View(reviews);
         }
