+++ Webserver/Models/ViewModels/Account/ExternalLoginListViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Webserver.Models.ViewModels.Account
+{
+    public class ExternalLoginListViewModel
+    {
+        public string ReturnUrl { get; set; }
+    }
+}
