+++ Webserver/Models/ViewModels/Account/SendCodeViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System.Collections.Generic;
+
+namespace Webserver.Models.ViewModels.Account
+{
+    public class SendCodeViewModel
+    {
+        public string SelectedProvider { get; set; }
+        public ICollection<System.Web.Mvc.SelectListItem> Providers { get; set; }
+        public string ReturnUrl { get; set; }
+        public bool RememberMe { get; set; }
+    }
+}
