+++ Webserver/Models/ViewModels/Manage/ConfigureTwoFactorViewModel.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Webserver.Models.ViewModels.Manage
+{
+    public class ConfigureTwoFactorViewModel
+    {
+        public string SelectedProvider { get; set; }
+        public ICollection<System.Web.Mvc.SelectListItem> Providers { get; set; }
+    }
+}
