+++ Webserver/Models/ViewModels/Manage/FactorViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace Webserver.Models.ViewModels.Manage
+{
+    public class FactorViewModel
+    {
+        public string Purpose { get; set; }
+    }
+}
