+++ Webserver/Models/ViewModels/Manage/IndexViewModel.cs
@@ -1,0 +1,14 @@
+﻿using Microsoft.AspNet.Identity;
+using System.Collections.Generic;
+
+namespace Webserver.Models.ViewModels.Manage
+{
+    public class IndexViewModel
+    {
+        public bool HasPassword { get; set; }
+        public IList<UserLoginInfo> Logins { get; set; }
+        public string PhoneNumber { get; set; }
+        public bool TwoFactor { get; set; }
+        public bool BrowserRemembered { get; set; }
+    }
+}
