+++ Webserver/Models/ViewModels/Manage/ManageLoginsViewModel.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNet.Identity;
+using Microsoft.Owin.Security;
+using System.Collections.Generic;
+
+namespace Webserver.Models.ViewModels.Manage
+{
+    public class ManageLoginsViewModel
+    {
+        public IList<UserLoginInfo> CurrentLogins { get; set; }
+        public IList<AuthenticationDescription> OtherLogins { get; set; }
+    }
+}
